move client constructor to multicluster package
