+++ pkg/multicluster/client.go
@@ -1,9 +1,30 @@
 package multicluster
 
-import "sigs.k8s.io/controller-runtime/pkg/client"
+import (
+	"github.com/rotisserie/eris"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
 
 // Client exposes client.Client for multiple clusters.
 type Client interface {
 	// Cluster returns a client.Client for the given cluster.
 	Cluster(name string) (client.Client, error)
+}
+
+type mcClient struct {
+	managers ManagerSet
+}
+
+var _ Client = &mcClient{}
+
+func NewClient(managers ManagerSet) *mcClient {
+	return &mcClient{managers: managers}
+}
+
+func (c *mcClient) Cluster(name string) (client.Client, error) {
+	mgr, err := c.managers.Cluster(name)
+	if err != nil {
+		return nil, eris.Wrapf(err, "failed to get client for cluster %v", name)
+	}
+	return mgr.GetClient(), nil
 }
